seems to be basically working, now its time to do some more serious tests
